Added some basic shape inheritance without visualization
Played arround with inheritance in and out of the class, need to change
the use of chars for the use of strings.

Also added a debug version to the Makefile. Also modified [make all] so
it cleans previous binaries and LLVM_IR files and runs the binary
compilation and LLVM_IR generation.